+++ src/bacnet/bacapp.h
@@ -16,6 +16,8 @@
 #include "bacnet/bacdef.h"
 /* BACnet Stack API */
 #include "bacnet/access_rule.h"
+#include "bacnet/authentication_factor.h"
+#include "bacnet/authentication_factor_format.h"
 #include "bacnet/bacaction.h"
 #include "bacnet/bacaddr.h"
 #include "bacnet/bacdest.h"
@@ -185,6 +187,10 @@ typedef struct BACnet_Application_Data_Value {
         BACNET_SC_HUB_FUNCTION_CONNECTION_STATUS SC_Hub_Function_Status;
         BACNET_SC_DIRECT_CONNECTION_STATUS SC_Direct_Status;
         BACNET_SC_HUB_CONNECTION_STATUS SC_Hub_Status;
+#endif
+#if defined(BACAPP_AUTHENTICATION)
+        BACNET_AUTHENTICATION_FACTOR_FORMAT Authentication_Format;
+        BACNET_AUTHENTICATION_FACTOR Authentication_Factor;
 #endif
     } type;
     /* simple linked list if needed */