+++ src/bacnet/bactext.h
@@ -385,6 +385,10 @@ bactext_access_event_name_default(uint32_t index, const char *default_string);
 
 BACNET_STACK_EXPORT
 const char *bactext_authentication_status_name(uint32_t index);
+BACNET_STACK_EXPORT
+const char *bactext_authentication_disable_reason_name(uint32_t index);
+BACNET_STACK_EXPORT
+const char *bactext_authentication_factor_type_name(uint32_t index);
 
 BACNET_STACK_EXPORT
 const char *bactext_authorization_mode_name(uint32_t index);